Fix support for system tray icons
Patches all application icons to use the application id as a
prefix, so that they are exported by flatpak-builder.

This ensures tray icons work as expected. It also removes the
need to rename icons in the manifest.

The new patch should be also upstreamable, assuming there are no issues
with other platforms supported by keepassxc (Windows and MacOS).

Fixes #5