keepassxc: Fix host access for attached files
Files attached to the KeePassXC database can be opened from within the
application. To enable this files are temporarily placed under /tmp.
Because this location is unavailable to the host, it fails.

Allowing filesystem access to /tmp, solves this. However, it would
probably be better if KeePassXC followed XDG conventions by using
XDG_RUNTIME_DIR.

Even better, in the case of flatpak it would be probably be best to use
XDG_RUNTIME_DIR/app/APPID. This could simply be set with the build time
option: `-DKEEPASSXC_BUILD_TYPE=Flatpak`

Fixes #14