Fix KeePassXC-Browser support
The KeePassXC-Browser extension did not work with Flatpak. The reasons
for this are many and there is no perfect solution. Still this should
fix it for most users. In cases where it does not, it is fixable with
minor configuration.

There are some limitations and pre-requisites for this to work:

1. Supported browsers (Chromium, Google Chrome, Firefox, Vivaldi) cannot
   sandbox their respective configuration directories. Host access must
   be possible.

2. This works by default, assuming the default system-wide installation
   is used. Specifically "/var/lib/flatpak/exports/bin" must contain
   "org.keepassxc.KeePassXC". If this is not the case users can override
   this by setting a custom proxy path within KeePassXC.