Grant direct access to host file system
Ideally this should not be necessary; it is the wrong solution. But
KeePassXC simply is not designed with portals in mind and the situation
is not improving, quite the opposite. Nor do I have the time and energy
to debug this anymore.

This is unlikely to change until portals are better supported and pushed
for with snaps.

Users should now find everything file access related (90% of all
issues) working as expected.

Since beta branches are now a thing on Flathub, portal testing will now
be available in the beta branch:
flatpak install flathub org.keepassxc.KeePassXC//beta