libargon2: Update to 20190702
The argon2 Makefile now assumes all Linux x86_64 builds target
Debian/Ubuntu and sets the library path to lib/x86_64-linux-gnu
instead of the old sensible & neutral default: "lib".

In a Flatpak this separation is obviously completely redundant, but this
also causes KeePassXC to fail to find the library during runtime.

The way to override this (also documented as comments in the Makefile)
is to override the LIBRARY_REL variable, using make arguments.