asciidoc: Remove build time dependency
The latest sdk (5.13) has removed python2 since it will soon reach
end-of-life. This breaks asciidoc, since it still depends on python2.
However both asciidoc and any man pages it is used to build are all
cleaned up, so it is ultimately not required to begin with.

By patching in an easy way to disable documentation for all affected
modules it is possible to remove asciidoc, solving the problem and
lowering build times.