Disable FreeDesktop Secrets service functionality
Cannot be guaranteed to work by default, due to unresolvable conflicts
with gnome-keyring.

Gnome Keyring includes a org.freedesktop.secrets D-Bus service that
allow dependent apps to launch it, if it isn't running. Depending on
the distribution or desktop environment it can be very difficult or
unpractical to disable this. Uninstalling it may also not be an option
due to package dependencies.

- From a trust perspective (not to mention Flathub policy?) it might be
  questionable to allow the app to own org.freedesktop.secrets.

- Future xdg portals, will depend on a sandbox compatible implementation
  of the service and because of this gnome-keyring could become a
  xdg-desktop-portal dependency.
  See: https://github.com/flatpak/xdg-desktop-portal/issues/311
       https://github.com/flatpak/xdg-desktop-portal/pull/359

- Any sandbox compatible implementation replacing gnome-keyring would
  surely need to run on its own, outside the application sandbox.

- Removing gnome-keyring and replacing it with an incomplete
  replacement could break other Flatpak apps in the future.