Add KStatusNotifierItem DBus ownership
KDE, Unity and other environments that support StatusNotifierItem
(KStatusNotifierItem), requires the app to register (own) the DBus
service `org.kde.StatusNotifierItem-X-Y` where X is the pid and Y is an
arbitrary unique numeric identifier between instances set by the same
application. The KDE standard seems based upon the freedesktop one [1].

Because this standard cannot guarantee a functional uid in namespaced
environments conflicts might occur with other flatpak apps;
specifically ones that allow ownership of the same service. In such
scenarios the tray icon will unavoidably remain broken.

This might also break upon updates if ex. the app changes its numeric
identifier. Ideally all, or at least app and runtime, updates should
now include be tested with this in mind.

The only alternative is `--own-name=org.kde.*` which not only increases
the risk of conflicts, but is a really broad and insecure permission.
It was dismissed for the same reasons as with the Riot app [3].

Notes:
- The workaround was tested using the: KStatusNotifierItem/AppIndicator
  GNOME extension [2].
- The numeric instance identifier: Y (or 2) should be stable and the
  fact that KeePassXC only allows a single instance helps a great deal.

[1] https://freedesktop.org/wiki/Specifications/StatusNotifierItem/StatusNotifierItem/
[2] https://extensions.gnome.org/extension/615/appindicator-support/
[3] https://github.com/flathub/im.riot.Riot/issues/100#issuecomment-625796916