Use keepassxc-wrapper in desktop entry file
By setting keepassxc-wrapper as the command in the manifest file
browsers are able to communicate with keepassxc-proxy. However this
does not affect DBus activation or the desktop entry file, which will
use keepassxc directly.

For the browser integration this does not matter; it works regardless.
But this does confuse things when the namespaced pid is changes between
2 & 3. Suddenly indicator support requires ownership of either:
org.kde.StatusNotifierItem-2-2 or org.kde.StatusNotifierItem-3-2

By making the behaviour consistent it is less confusing and hopefully
this removes the need to also allow:
`--own-name=org.kde.StatusNotifierItem-2-2`
which most likely is even more prone to conflicts.