Fix browser integration for old flatpak releases
Users of CentOS only have flatpak 1.0.9 available and that does not
include the commit that exports `FLATPAK_ID` with flatpak-run(1).
Consequently browser integration broke because the socket was not
listening at the correct location.

In retrospect it seems stupid to have relied upon FLATPAK_ID and so
the solution is to revert to hard coding the id wherever it needed.

Fixes #67