Add second NotifierItem permission
KeePassXC now seem to require a second registered StatusNotifierItem.
The first one will still be using 2-2 as the id and if that doesn't work
it will try again with an increment of 3-2 and so on. A second service
will use the increment above that.

It us unknown what it is used for but it might as well be allowed.